Porting file-share and file-datalake perf tests to the new model
