+++ sdk/storage/perf-tests/storage-file-datalake/sample.env
@@ -1,0 +1,1 @@
+STORAGE_CONNECTION_STRING=<connection-string>
