+++ sdk/storage/perf-tests/storage-file-datalake/test/append.spec.ts
@@ -3,8 +3,8 @@
 
 import { PerfStressOptionDictionary } from "@azure/test-utils-perfstress";
 import { StorageDFSTest } from "./storageTest.spec";
-import { DataLakeFileClient } from "../../../src";
-import { generateUuid } from "@azure/core-http";
+import { DataLakeFileClient } from "@azure/storage-file-datalake";
+import { v4 as generateUuid } from "uuid";
 interface StorageDFSAppendTestOptions {
   size: number;
 }
@@ -17,8 +17,8 @@ export class StorageDFSAppendTest extends StorageDFSTest<StorageDFSAppendTestOpt
       description: "Size to append in bytes",
       shortName: "sz",
       longName: "size",
-      defaultValue: 1024
-    }
+      defaultValue: 1024,
+    },
   };
   fileClient: DataLakeFileClient;
