+++ sdk/storage/perf-tests/storage-file-datalake/test/storageTest.spec.ts
@@ -7,12 +7,12 @@ import {
   DataLakeServiceClient,
   StorageSharedKeyCredential,
   DataLakeFileSystemClient,
-  DataLakeDirectoryClient
-} from "../../../src";
+  DataLakeDirectoryClient,
+} from "@azure/storage-file-datalake";
 
 // Expects the .env file at the same level as the "test" folder
 import * as dotenv from "dotenv";
-import { generateUuid } from "@azure/core-http";
+import { v4 as generateUuid } from "uuid";
 dotenv.config();
 
 export abstract class StorageDFSTest<TOptions> extends PerfStressTest<TOptions> {
