+++ sdk/storage/perf-tests/storage-file-datalake/test/upload.spec.ts
@@ -1,37 +1,36 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { generateUuid } from "@azure/core-http";
+import { v4 as generateUuid } from "uuid";
 import { PerfStressOptionDictionary } from "@azure/test-utils-perfstress";
-import { ShareFileClient } from "../../../src";
-import { StorageFileShareTest } from "./storageTest.spec";
+import { DataLakeFileClient } from "@azure/storage-file-datalake";
+import { StorageDFSTest } from "./storageTest.spec";
 
 interface StorageFileShareUploadTestOptions {
   size: number;
 }
 
-export class StorageFileShareUploadTest extends StorageFileShareTest<
-  StorageFileShareUploadTestOptions
-> {
-  fileClient: ShareFileClient;
-  buffer = Buffer.alloc(this.parsedOptions.size.value!);
+export class StorageDFSUploadTest extends StorageDFSTest<StorageFileShareUploadTestOptions> {
+  buffer: Buffer;
+  fileClient: DataLakeFileClient;
   public options: PerfStressOptionDictionary<StorageFileShareUploadTestOptions> = {
     size: {
       required: true,
       description: "Size in bytes",
       shortName: "sz",
       longName: "size",
-      defaultValue: 1024
-    }
+      defaultValue: 1024,
+    },
   };
 
   constructor() {
     super();
     const fileName = generateUuid();
     this.fileClient = this.directoryClient.getFileClient(fileName);
+    this.buffer = Buffer.alloc(this.parsedOptions.size.value!);
   }
 
   async runAsync(): Promise<void> {
-    await this.fileClient.uploadData(this.buffer);
+    await this.fileClient.upload(this.buffer);
   }
 }
