+++ sdk/storage/perf-tests/storage-file-share-track-1/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@azure-tests/perf-storage-blob",
+  "name": "@azure-tests/perf-storage-file-share-track-1",
   "version": "1.0.0",
   "description": "",
   "main": "",
@@ -7,15 +7,11 @@
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@azure/core-http": "^1.2.0",
-    "@azure/core-https": "1.0.0-beta.2",
-    "@azure/storage-blob": "^12.5.0-beta.1",
-    "@azure/test-utils-perfstress": "^1.0.0",
+    "@azure/storage-file": "^10.3.0",
+    "@azure/test-utils-perfstress": "file:../../../test-utils/perfstress/azure-test-utils-perfstress-1.0.0.tgz",
     "@types/uuid": "^8.0.0",
     "@types/node": "^8.0.0",
-    "@types/node-fetch": "^2.5.0",
     "dotenv": "^8.2.0",
-    "node-fetch": "^2.6.0",
     "uuid": "^8.3.0",
     "tslib": "^2.0.0",
     "ts-node": "^8.3.0",
@@ -35,9 +31,10 @@
     "integration-test:node": "echo skipped",
     "integration-test": "echo skipped",
     "lint:fix": "eslint package.json src test --ext .ts --fix --fix-type [problem,suggestion]",
-    "lint": "eslint package.json src test --ext .ts -f html -o storage-blob-perf-test-lintReport.html || exit 0",
+    "lint": "eslint package.json src test --ext .ts -f html -o storage-file-share-perf-test-lintReport.html || exit 0",
     "pack": "npm pack 2>&1",
     "prebuild": "npm run clean",
+    "setup": "node ../../../../common/tools/perf-tests-track-1-setup.js",
     "unit-test:browser": "echo skipped",
     "unit-test:node": "echo skipped",
     "unit-test": "echo skipped",
