+++ sdk/storage/perf-tests/storage-file-share-track-1/sample.env
@@ -1,0 +1,1 @@
+STORAGE_CONNECTION_STRING=<connection-string>
