+++ sdk/storage/perf-tests/storage-file-share-track-1/test/upload.spec.ts
@@ -9,25 +9,24 @@ import { generateUuid } from "@azure/core-http";
 interface StorageFileShareUploadTestOptions {
   size: number;
 }
-export class StorageFileShareUploadTest extends StorageFileShareTest<
-  StorageFileShareUploadTestOptions
-> {
-  buffer = Buffer.alloc(this.parsedOptions.size.value!);
+export class StorageFileShareUploadTest extends StorageFileShareTest<StorageFileShareUploadTestOptions> {
   fileClient: FileURL;
+  buffer: Buffer;
   public options: PerfStressOptionDictionary<StorageFileShareUploadTestOptions> = {
     size: {
       required: true,
       description: "Size in bytes",
       shortName: "sz",
       longName: "size",
-      defaultValue: 1024
-    }
+      defaultValue: 1024,
+    },
   };
 
   constructor() {
     super();
     const fileName = generateUuid();
     this.fileClient = FileURL.fromDirectoryURL(this.directoryClient, fileName);
+    this.buffer = Buffer.alloc(this.parsedOptions.size.value!);
   }
 
   async setup() {
