+++ sdk/storage/perf-tests/storage-file-share/test/download.spec.ts
@@ -1,26 +1,24 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { generateUuid } from "@azure/core-http";
+import { v4 as generateUuid } from "uuid";
 import { PerfStressOptionDictionary, drainStream } from "@azure/test-utils-perfstress";
-import { ShareFileClient } from "../../../src";
+import { ShareFileClient } from "@azure/storage-file-share";
 
 import { StorageFileShareTest } from "./storageTest.spec";
 interface StorageFileShareDownloadTestOptions {
   size: number;
 }
 
-export class StorageFileShareDownloadTest extends StorageFileShareTest<
-  StorageFileShareDownloadTestOptions
-> {
+export class StorageFileShareDownloadTest extends StorageFileShareTest<StorageFileShareDownloadTestOptions> {
   public options: PerfStressOptionDictionary<StorageFileShareDownloadTestOptions> = {
     size: {
       required: true,
       description: "Size in bytes",
       shortName: "sz",
       longName: "size",
-      defaultValue: 1024
-    }
+      defaultValue: 1024,
+    },
   };
   static fileName = generateUuid();
   fileClient: ShareFileClient;
