+++ sdk/storage/perf-tests/storage-file-share/test/upload.spec.ts
@@ -1,26 +1,26 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-import { generateUuid } from "@azure/core-http";
+import { v4 as generateUuid } from "uuid";
 import { PerfStressOptionDictionary } from "@azure/test-utils-perfstress";
-import { DataLakeFileClient } from "../../../src";
-import { StorageDFSTest } from "./storageTest.spec";
+import { ShareFileClient } from "@azure/storage-file-share";
+import { StorageFileShareTest } from "./storageTest.spec";
 
 interface StorageFileShareUploadTestOptions {
   size: number;
 }
 
-export class StorageDFSUploadTest extends StorageDFSTest<StorageFileShareUploadTestOptions> {
+export class StorageFileShareUploadTest extends StorageFileShareTest<StorageFileShareUploadTestOptions> {
+  fileClient: ShareFileClient;
   buffer: Buffer;
-  fileClient: DataLakeFileClient;
   public options: PerfStressOptionDictionary<StorageFileShareUploadTestOptions> = {
     size: {
       required: true,
       description: "Size in bytes",
       shortName: "sz",
       longName: "size",
-      defaultValue: 1024
-    }
+      defaultValue: 1024,
+    },
   };
 
   constructor() {
@@ -31,6 +31,6 @@ export class StorageDFSUploadTest extends StorageDFSTest<StorageFileShareUploadT
   }
 
   async runAsync(): Promise<void> {
-    await this.fileClient.upload(this.buffer);
+    await this.fileClient.uploadData(this.buffer);
   }
 }
