--- sdk/storage/storage-file-share/test/perfstress/track-1/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "compilerOptions": {
-    "module": "commonjs"
-  }
-}
